+++ dashboard-spring/src/main/java/com/humintecTest/dashboard/vo/PidVo.java
@@ -7,8 +7,8 @@ import java.sql.Date;
 @Data
 public class PidVo {
     private Date last_scan_date;
-    private  String pathID_ch;
-    private int amount;
+    private String pathID_ch;
+    private long amount;
     private String start_date;
     private String end_date;
 }
